+++ src/router/index.js
@@ -15,7 +15,7 @@ import ReviewBoardView from '../components/post/review/ReviewBoardView.vue';
 import RegistReviewPostView from '../components/post/review/RegistReviewPostView.vue';
 import ReviewPostView from '../components/post/review/ReviewPostView.vue';
 import FashionBoardView from '../components/post/fashion/FashionBoardView.vue';
-import FashionPostView from '../components/FashionPostView.vue';
+import FashionPostView from '../components/post/fashion/FashionPostView.vue';
 
 // Placeholders/Imports from features branch (Uncomment and complete as needed)
 // import FashionPostView from '../components/FashionPostView.vue'; // Likely replaced by FashionBoardView or similar
